requirements.txt
biomaj
biomaj-user
biomaj-zipkin
redis
PyYAML
flask
python-consul
prometheus_client
requests
